requirements.txt
# --- Core API ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# --- LLM Providers ---
httpx>=0.25.0
google-genai>=1.0.0

# --- Image Processing ---
Pillow>=10.0.0
cairosvg>=2.7.0

# --- Document Parsing ---
PyMuPDF>=1.23.0
lxml>=4.9.0

# --- Archive Support ---
py7zr>=0.20.0
rarfile>=4.1
patool>=1.12

# --- OCR (requires tesseract binary) ---
pytesseract>=0.3.10

# --- CLI & Utilities ---
click>=8.1.0
tqdm>=4.66.0
rich>=13.0.0

# --- Async ---
aiofiles>=23.2.0
